tests: Added unit test for deworming service

+++ api/src/modules/deworming/deworming.service.js
@@ -1,4 +1,5 @@
 const Boom = require('@hapi/boom');
+const { Op } = require('sequelize');
 const { v4: uuidv4 } = require('uuid');
 
 const dewormingRepository = require('./deworming.repository');
@@ -9,49 +10,84 @@ const formatDeworming = (dewormer) => {
     id: dewormer.id,
     dewormer: dewormer.dewormer,
     description: dewormer.description || null,
-    animalId: dewormer.animal ? dewormer.animal.code : null,
+    animal: dewormer.animal ? dewormer.animal.code : null,
     registeredAt:
       dewormer.registeredAt.toISOString().split('T')[0] ||
       dewormer.registeredAt,
   };
 };
 
-const getAllDeworming = async (userId, filters) => {
+const buildFilters = (query) => {
+  const { dewormer, animalId } = query;
+
+  return {
+    ...(dewormer && { dewormer: { [Op.iLike]: `%${dewormer}%` } }),
+    ...(animalId && { animalId }),
+  };
+};
+
+const getAllDeworming = async (userId, query) => {
+  if (!query || typeof query !== 'object') {
+    throw Boom.badRequest('query was not provided');
+  }
+
+  const filters = buildFilters(query);
+
   const deworming = await dewormingRepository.findAllDewormings(
     userId,
     filters,
   );
 
+  if (deworming?.length === 0) return [];
+
   return deworming.map((d) => formatDeworming(d));
 };
 
 const getDeworming = async (userId, dewormingId) => {
+  if (!userId) throw Boom.badRequest('userId was not provided');
+  if (!dewormingId) throw Boom.badRequest('dewormingId was not provided');
+
   const deworming = await dewormingRepository.findOne(userId, dewormingId);
   if (!deworming?.id) throw Boom.notFound('Deworming does not exist');
+
   return formatDeworming(deworming);
 };
 
 const createDeworming = async (userId, dewormingData) => {
+  if (!userId) throw Boom.badRequest('userId was not provided');
+  if (!dewormingData?.dewormer || !dewormingData?.animalId) {
+    throw Boom.badRequest('dewormer was not provided');
+  }
+
   const animal = await animalRepository.findOne(userId, dewormingData.animalId);
-  if (!animal?.id)
+  if (!animal?.id) {
     throw Boom.notFound('Animal does not exist or does not belong to the user');
+  }
 
   const deworming = {
     id: uuidv4(),
     dewormer: dewormingData.dewormer,
-    description: dewormingData.description || null,
+    description: dewormingData?.description || null,
     animalId: dewormingData.animalId,
     registeredAt:
       dewormingData.registeredAt || new Date().toISOString().split('T')[0],
   };
 
   const newdeworming = await dewormingRepository.create(deworming);
-  if (!newdeworming?.id)
+  if (!newdeworming?.id) {
     throw Boom.badRequest('Something went wrong creating the deworming');
+  }
+
   return newdeworming;
 };
 
 const updateDeworming = async (userId, dewormingId, dewormingData) => {
+  if (!userId) throw Boom.badRequest('userId was not provided');
+  if (!dewormingId) throw Boom.badRequest('dewormingId was not provided');
+  if (!dewormingData?.dewormer && !dewormingData.description) {
+    throw Boom.badRequest('dewormingData was not provided');
+  }
+
   const deworming = await dewormingRepository.findOne(userId, dewormingId);
   if (!deworming?.id) throw Boom.conflict('deworming does not exists');
 
@@ -73,10 +109,19 @@ const updateDeworming = async (userId, dewormingId, dewormingData) => {
 };
 
 const deleteDeworming = async (userId, dewormingId) => {
+  if (!userId) throw Boom.badRequest('userId was not provided');
+  if (!dewormingId) throw Boom.badRequest('dewormingId was not provided');
+
   const deworming = await dewormingRepository.findOne(userId, dewormingId);
   if (!deworming?.id) throw Boom.conflict('deworming does not exists');
 
-  return dewormingRepository.destroy(dewormingId);
+  const affectedRows = await dewormingRepository.destroy(dewormingId);
+
+  if (affectedRows === 0) {
+    throw Boom.badRequest('Something went wrong deleting the deworming');
+  }
+
+  return affectedRows;
 };
 
 module.exports = {
